# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Read-only catalog API for the eFootball marketplace
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductStatus = typeof ProductStatus[keyof typeof ProductStatus];


export const ProductStatus = {
  Available: 'Available',
  Reserved: 'Reserved',
  Sold: 'Sold',
  Hidden: 'Hidden',
} as const;

export interface Product {
  id: string;
  title: string;
  description: string;
  image: string;
  coins: number;
  gp: number;
  price: number;
  /** @nullable */
  originalPrice: number | null;
  category: string;
  platform: string;
  status: ProductStatus;
  productIdentifier: string;
  deliveryMethod: string;
  featured: boolean;
  /** @nullable */
  offerId: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface Category {
  id: string;
  name: string;
  slug: string;
  createdAt: string;
}

export interface Offer {
  id: string;
  title: string;
  description: string;
  originalPrice: number;
  discountedPrice: number;
  discountPercentage: number;
  image: string;
  startDate: string;
  endDate: string;
  active: boolean;
  featured: boolean;
}

export interface Error {
  error: string;
}

export type SearchParameter = string;

export type CategoryParameter = string;

export type PlatformParameter = string;

export type StatusParameter = ProductStatus;

export type MinCoinsParameter = number;

export type MaxCoinsParameter = number;

export type MinGpParameter = number;

export type MaxGpParameter = number;

export type MinPriceParameter = number;

export type MaxPriceParameter = number;

export type FeaturedParameter = boolean;

export type OffersParameter = boolean;

export type SortParameter = typeof SortParameter[keyof typeof SortParameter];


export const SortParameter = {
  newest: 'newest',
  oldest: 'oldest',
  'price-low': 'price-low',
  'price-high': 'price-high',
  'coins-low': 'coins-low',
  'coins-high': 'coins-high',
  'gp-low': 'gp-low',
  'gp-high': 'gp-high',
} as const;

export type ListProductsParams = {
search?: SearchParameter;
category?: CategoryParameter;
platform?: PlatformParameter;
status?: StatusParameter;
/**
 * @minimum 0
 */
minCoins?: MinCoinsParameter;
/**
 * @minimum 0
 */
maxCoins?: MaxCoinsParameter;
/**
 * @minimum 0
 */
minGp?: MinGpParameter;
/**
 * @minimum 0
 */
maxGp?: MaxGpParameter;
/**
 * @minimum 0
 */
minPrice?: MinPriceParameter;
/**
 * @minimum 0
 */
maxPrice?: MaxPriceParameter;
featured?: FeaturedParameter;
offers?: OffersParameter;
sort?: SortParameter;
};

export type ListOffersParams = {
active?: boolean;
featured?: boolean;
};