# lib/api-zod/src/generated/types/offer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Read-only catalog API for the eFootball marketplace
 * OpenAPI spec version: 0.1.0
 */

export interface Offer {
  id: string;
  title: string;
  description: string;
  originalPrice: number;
  discountedPrice: number;
  discountPercentage: number;
  image: string;
  startDate: Date;
  endDate: Date;
  active: boolean;
  featured: boolean;
}