# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Read-only catalog API for the eFootball marketplace
 * OpenAPI spec version: 0.1.0
 */
import type { ProductStatus } from './productStatus';

export interface Product {
  id: string;
  title: string;
  description: string;
  image: string;
  coins: number;
  gp: number;
  price: number;
  /** @nullable */
  originalPrice: number | null;
  category: string;
  platform: string;
  status: ProductStatus;
  productIdentifier: string;
  deliveryMethod: string;
  featured: boolean;
  /** @nullable */
  offerId: string | null;
  createdAt: Date;
  updatedAt: Date;
}